+++ database/seeders/DemoSeeder.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Bithoven\Dummy\Database\Seeders;
+
+use Illuminate\Database\Seeder;
+
+class DemoSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     */
+    public function run(): void
+    {
+        $this->call([
+            DummyDemoSeeder::class,
+        ]);
+    }
+}